1, update with IAMF PR:471. 2, fix asan detection issue. 3, remove duplicated wav write/read files

+++ code/test/tools/iamfplayer/include/a2b_endian.h
@@ -58,7 +58,7 @@ static inline uint16_t bswap16(const uint16_t u16) {
     ((__GNUC__ > 4) || ((__GNUC__ == 4) && (__GNUC_MINOR__ >= 8)))
   return __builtin_bswap16(u16);
 #else
-  return (u16 << 8) | (u16 >> 8);
+  return (uint16_t)((u16 << 8) | (u16 >> 8));
 #endif
 #else
   return u16;
